requirements.txt
Flask==2.3.3
Flask-APScheduler==1.12.4
Flask-Migrate==4.0.4
Flask-Login==0.6.2
Flask_Reuploaded==1.3.0
Flask-SQLAlchemy==3.0.5
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
PyMySQL==1.1.0
psutil==5.9.5
Flask-Mail==0.9.1
Pillow==10.0.0
cryptography==41.0.3
validators==0.22.0
Werkzeug==2.3.7

numpy==1.24.3
pandas
scikit-learn
openpyxl

openai==0.27.4
tiktoken

websockets==10.4